requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
numpy==1.26.4
soundfile==0.12.1
librosa==0.10.2.post1
faster-whisper==1.1.0
allosaurus==1.0.2
panphon==0.20.0
g2p_en==2.1.0
pronouncing==0.2.0
g2p-en-plus==2.0.3
phonemizer==3.3.0
transformers==4.43.3
torch==2.6.0
montreal-forced-aligner==3.3.8
textgrid==1.6.1
# MFA requires downloaded acoustic + dictionary models (see docs)
safetensors==0.4.4
huggingface_hub==0.24.5
accelerate==0.33.0
openai==1.54.3
python-dotenv==1.0.0
# Database and API key management
supabase==2.9.1
asyncpg==0.29.0
cryptography==43.0.1
PyJWT==2.9.0
psycopg[binary]==3.2.1
requests==2.32.3
boto3==1.35.78
